move 'nvm' extractor  to extractor/filesystem/runtime/nodejs/nvm/
