+++ extractor/filesystem/runtime/nodejs/nvm/testdata/simpleValidWithComments/.nvmrc
