+++ extractor/filesystem/runtime/nodejs/nvm/testdata/validMultiVersionWithSkip/.nvmrc
