+++ extractor/filesystem/runtime/nodejs/nvm/testdata/validWhiteSpaces/.nvmrc
